# contracts/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for the Reverse Marketplace platform",
    "title": "Reverse Marketplace API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/admin/announcements/": {
      "post": {
        "description": "Create an announcement banner",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "endsAt": {
                    "description": "When the banner expires. Null/omitted = open-ended.",
                    "nullable": true
                  },
                  "message": {
                    "description": "Banner text shown to users (follow RUNBOOK_OPS.md §2 wording rules)",
                    "maxLength": 500,
                    "minLength": 3,
                    "type": "string"
                  },
                  "severity": {
                    "default": "info",
                    "description": "Banner severity — info (purple), warning (amber), critical (red)",
                    "enum": [
                      "info",
                      "warning",
                      "critical"
                    ],
                    "type": "string"
                  },
                  "startsAt": {
                    "description": "When the banner becomes active. Defaults to now."
                  }
                },
                "required": [
                  "message"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/announcements/{id}": {
      "delete": {
        "description": "Clear an announcement banner",
        "parameters": [
          {
            "description": "Announcement UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Reverse Marketplace API (base path)",
      "url": "/api/v1"
    }
  ]
}